+++ CHANGELOG.md
@@ -12,6 +12,14 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+## [1.0.1.19] - 2020-07-02
+
+### Changed
+
+* function Update-DSM7SoftwareCategory
+* function New-DSM7SoftwareCategory
+
+
 ## [1.0.1.18] - 2020-06-09
 
 ### Added